Performance tuning day 16

+++ src/main/java/com/adventofcode/common/InputHelper.java
@@ -40,8 +40,8 @@ public class InputHelper {
         return textList.stream().map(Integer::parseInt).toList();
     }
 
-    public List<Integer> parseStringToSingleDigitIntList(String text) {
-        return Arrays.stream(text.split("")).map(Integer::parseInt).toList();
+    public List<Byte> parseStringToSingleDigitIntList(String text) {
+        return Arrays.stream(text.split("")).map(Byte::parseByte).toList();
     }
 
 }
